Add a stub to properly load nouveau
Since nvidia-utils installs files in ``/usr/lib/modprobe.d/`` we have
the right to completely override their behavior with files with the same
name in ``/etc/modprobe.d``. This commit should fix the issue with the
unloadable nouveau module for LiveCD. Note that this config should not
be copied into the installation system.